fireball attack, damage flinch, fireball deflect.
I refined the fireball attack. There were some inconsistencies with how the fireballs were being launched which would make it difficult for the players to really predict where the projectiles would launch. Played with different firing patterns. Implemented the deflect mechanic and made sure it worked reliably. Noticed the fireball would stop bouncing after a certain number of bounces. Added a physic material to the ball and deflection to reduce the friction to zero. I wired up the damage flinch animation and integrated it into the command pattern I'm using for the dragon actions. confirmed the flinch animation would trigger as well as restrict dragon movement and rotation during the flinch.